Updated script to accept 2 feeds and merge based on newest first
This script accepts two rss feed urls and filter them to be only the feed updates posted in the last rolling 24-period. it then sorts them by oldest post first and outputs a list of the feed updates to the command line. It also creates a csv file of the same information